+++ database/migrations/2016_10_22_105310_create_locations_table.php
@@ -14,7 +14,7 @@ class CreateLocationsTable extends Migration
     public function up()
     {
         Schema::create('locations', function (Blueprint $table) {
-            $table->unsignedInteger('location_id')->unique();
+            $table->unsignedBigInteger('location_id')->unique();
             $table->string('location_name');
             $table->timestamps();
         });
