+++ database/migrations/2016_10_22_105335_create_musics_table.php
@@ -14,7 +14,7 @@ class CreateMusicsTable extends Migration
     public function up()
     {
         Schema::create('musics', function (Blueprint $table) {
-            $table->unsignedInteger('music_id')->unique();
+            $table->unsignedBigInteger('music_id')->unique();
             $table->string('music_name');
             $table->timestamps();
         });
