+++ database/migrations/2016_10_22_105359_create_profiles_friends_table.php
@@ -14,8 +14,8 @@ class CreateProfilesFriendsTable extends Migration
     public function up()
     {
         Schema::create('profiles_friends', function (Blueprint $table) {
-            $table->unsignedInteger('profile_id');
-            $table->unsignedInteger('friend_profile_id');
+            $table->unsignedBigInteger('profile_id');
+            $table->unsignedBigInteger('friend_profile_id');
             $table->unique(array('profile_id','friend_profile_id'));
             $table->timestamps();
         });
