+++ database/migrations/2016_10_22_105407_create_profiles_hometowns_table.php
@@ -14,8 +14,8 @@ class CreateProfilesHometownsTable extends Migration
     public function up()
     {
         Schema::create('profiles_hometowns', function (Blueprint $table) {
-            $table->unsignedInteger('profile_id');
-            $table->unsignedInteger('hometown_id');
+            $table->unsignedBigInteger('profile_id');
+            $table->unsignedBigInteger('hometown_id');
             $table->unique(array('profile_id','hometown_id'));
             $table->timestamps();
         });
